Move function to utils/sa.py
